spring-security (complete)

+++ src/main/java/org/example/surveyapi/web/dtos/response/UserResponse.java
@@ -10,5 +10,5 @@ public class UserResponse {
     private String name;
     private String lastName;
     private String email;
-    private Long phoneNumber;
+    private Integer phoneNumber;
 }
